finalchat/requirements.txt
pyasn1==0.5.0
pycryptodome==3.19.0
rsa==4.9
tqdm==4.66.1
